#163 - various
1. Add per label quick edit forms
2. split up updateCharacter and quickUpdateCharacter mutations to avoid generalization rabbit hole
