+++ api/seeds/10_users.js
@@ -3,10 +3,12 @@ exports.seed = function(knex, Promise) {
     .then(() =>
       knex('users').insert([
         {
-          auth0UserId: 'oauth2|twitch|177040760'  // mike twitch
+          auth0UserId: 'oauth2|twitch|177040760',  // mike twitch
+          name: 'Mike'
         },
         {
-          auth0UserId: 'google-oauth2|103015308867869178931'  // nazar gmail
+          auth0UserId: 'google-oauth2|103015308867869178931',  // nazar gmail
+          name: 'Nazar'
         }
       ]));
 };
