+++ api/src/engine/1_5e/index.js
@@ -15,5 +15,4 @@ export default class {
       stripUnknown: true
     });
   }
-
 }
