+++ client/src/components/Characters/components/CharacterLabelQuickEditForms/1_5e/index.js
@@ -1,0 +1,11 @@
+import CharacterLabelQuickEditForm from './1.0.0/CharacterLabelQuickEditForm.js';
+
+const versions = {
+  '1.0.0': {
+    CharacterLabelQuickEditForm
+  }
+};
+
+export function CharacterQuickEditForm(version) {
+  return versions[version].CharacterLabelQuickEditForm;
+}
